Address review feedback on focus fallback
