+++ src/config/toml.rs
@@ -129,7 +129,6 @@ pub(super) struct EdgePanConfig {
 pub(super) struct ZoomConfig {
     pub step: Option<f64>,
     pub fit_padding: Option<f64>,
-    pub navigation_target: Option<f64>,
     pub reset_on_new_window: Option<bool>,
     pub reset_on_activation: Option<bool>,
 }
